+++ lib/test.dart
@@ -21,6 +21,7 @@ export 'package:matcher/matcher.dart';
 export 'src/frontend/expect.dart';
 export 'src/frontend/expect_async.dart';
 export 'src/frontend/future_matchers.dart';
+export 'src/frontend/on_platform.dart';
 export 'src/frontend/prints_matcher.dart';
 export 'src/frontend/skip.dart';
 export 'src/frontend/test_on.dart';